northd: Allow backwards compatibility for Logical_Switch_Port.up.
In general, ovn-northd expects ovn-controller to set Port_Binding.up
before it declares the logical switch port as being up.

Even though the recommended upgrade procedure for OVN states that
ovn-controllers should be upgraded before ovn-northd, there are cases
when CMSs don't follow this guideline.

This would cause all existing and bound Logical_Switch_Ports to be
declared "down" until ovn-controllers are upgraded.

To avoid this situation, ovn-controllers now explicitly set
Chassis.other_config:port-up-notif in their own chassis record.  Based
on this value, ovn-northd can determine if it needs to use the old type
of logic or the new one (Port_Binding.up) when setting LSP.up.

Note:
In case of downgrading ovn-controller before ovn-northd, if
ovn-controller is forcefully stopped it will not clear its chassis
record from the SB. Older versions will not have the capability to
clear the other_config:port-up-notif value so LSPs will be declared
"down" until ovn-northd is downgraded as well.  As this
upgrade/downgrade procedure is not the recommended one, we don't try
to deal with this scenario.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>