Fix submodule build when using build directory.
The initial addition of OVS as a submodule made the assumption that the
build is being performed from the OVN source directory. However, some
devs create a separate build directory for OVN. This means the search
for the OVS source will fail.

This fixes the problem by explicitly stating that he ovs submodule is
located in the ovn source directory.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>