binding: Fix potential crash when binding_seqno_run is skipped.
The prerequisite for binding_seqno_run() is that it has to be called
while there are valid SB and OVS transactions available.

ovn-controller's main() function respects that but that means that there
is a small window (when OVS transaction is still in progress) when port
bindings might be deleted from the Southbound.

When binding_seqno_run() was finally called, it was asserting that the
Port_bindings to still exist.  This check is too restrictive and the
solution is to just skip bindings that have become partial (i.e., no
SB.Port_Binding or no OVS.Interface).

Fixes: 4d3cb42b076b ("binding: Set Logical_Switch_Port.up when all OVS flows are installed.")
Reported-at: https://bugzilla.redhat.com/1930030
Reported-by: Numan Siddique <numans@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>