controller: Add support for PTR DNS requests.
The native OVN DNS support doesn't yet support for PTR DNS requests.
This patch adds the support for it.  If suppose there is a dns record
as - "vm1.ovn.org"="10.0.0.4", then a normal DNS request will query for
"vm1.ovn.org" and the reply will be the IP address - 10.0.0.4.
PTR DNS request helps in getting the domain name of the IP address.
For the above example, the PTR DNS request will have a query name as
- "4.0.0.10.in-addr.arpa".  And the response will have "vm1.ovn.org".
In order to support this feature, this patch expects the CMS to define
an another entry in the DNS record  as - "4.0.0.10.in-addr.arpa"="vm1.ovn.org".

This makes the job of ovn-controller easier to support this feature.

Submitted-at: https://github.com/ovn-org/ovn/pull/74
Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: Numan Siddique <numans@ovn.org>