ovn-northd-ddlog: Use object form of is_some(), drop is_none().
DDlog supports a couple different forms of functions.  You can write
is_some(expr) or expr.is_some(), for example.  The latter is usually
easier to understand, especially in bigger expressions.  It was
introduced later so it's not consistently used in ovn-northd-ddlog.
This commit switches to the object-like form throughout the code.

Writing is_none(expr) or expr.is_none() is equivalent to writing
expr == None.  I think that the latter is easier to understand, so
this commit also makes that change throughout.

This code refactoring shouldn't change ovn-northd-ddlog behavior.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>