ovn-northd-ddlog: Define in_addr, in6_addr, eth_addr in ddlog code.
All of these were defined as opaque "extern" types in the ddlog code.
I have found that they are easier to work with and understand if
the ddlog code can look into their implementations.  This commit
re-defines them in terms of proper ddlog types.  It also renames
many of the functions that operate on them so that they can be used
in an object-like form, as well as defining many of these functions
in ddlog code rather than as externs.

This code refactoring shouldn't change ovn-northd-ddlog behavior.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Acked-by: Numan Siddique <numans@ovn.org>