northd: Optimize ct nat for load balancer traffic.
For a load balancer traffic destined to VIP, we do the below actions
related to conntrack in the ingress logical switch pipeline.
  1.  Send the packet to conntrack - ct()
  2a. if ct.new then ct_lb(backends)
  2b. if ct.est then ct(nat)

The step 2b is unnecessary and can be removed to avoid another
recirculation.

This patch improves this by doing the below.

For a load balancer traffic destined to VIP, we will now do
  1.   ct(nat)
  2a.  if ct.new - ct_lb(backends)
  2b.  if ct.est, no further action related to conntrack.

For non load balancer connection traffic, we will now do
  1.  ct(nat)
  2a. if ct.new then ct(commit)
  2b  if ct.est, no further action related to conntrack.

The same improvement is done for the egress logical switch
pipeline.  The stages - ls_in_lb and ls_out_lb are removed
since these stages are no longer needed.

Acked-by: Mark D. Gray <mark.d.gray@redhat.com>
Acked-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: Numan Siddique <numans@ovn.org>