Static Routes: Add ability to specify "discard" nexthop
Physical switches have the ability to specify "discard" or sometimes
"NULL interface" as a nexthop for routes. This can be used to prevent
routing loops in the network. Add a similar configuration for ovn
where nexthop accepts the string "discard". When the nexthop is discard
the action in the routing table will be to drop the packets.

Signed-off-by: Karthik Chandrashekar <karthik.c@nutanix.com>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>

----
v1 -> v2
----
* Add ddlog support
* Don't allow nexthop = "discard" when ECMP and BFD is set
----
v2 -> v3
----
* Use ddlog optimizations from review
Acked-by: Mark Michelson <mmichels@redhat.com>