ovn-nbctl: Don't replicate entire database unnecessarily.
ovn-nbctl was retrieving a replica of the entire northbound database
regardless of what it was asked to do.  This was unnecessary, but it
was not written to be smart enough to just retrieve what it needed.
This commit fixes the problem.  It makes "ovn-nbctl init", for example,
run instantly on a big database, whereas before it could take
arbitrarily long.

Signed-off-by: Ben Pfaff <blp@ovn.org>