tests: Fix frequent failure of "4 HV, 1 LS, 1 LR, packet test with HA distributed router gateway port:".
This test is failing quite often due to timing issues.  The failures are
mainly due to the packet not received by the other side of the tunnel.
The test case resets the pcap file and then injects the packet.  And
this packet gets lost sometimes if the pcap files are not reset yet.
To fix this issue the test now ensures that ovs-vswitchd resets the
pcap file before injecting the packet.  The test also now adds
some waits to make sure flows related to tunnels are programmed
when the chassisresident port moves from one gw chassis to other.

Signed-off-by: Numan Siddique <numans@ovn.org>
Acked-by: Mark Michelson <mmichels@redhat.com>