binding: Don't reset expected seqno for interfaces already being installed.
If an interface is already in 'binding_iface_seqno_map' then that means
that the openflow modifications needed for its functionality have
already been sent to ovs-vswitchd and we're waiting for an openflow
barrier reply.  In such cases, don't reset the expected seqno because
this would delay the notification unnecessarily.

This scenario can be hit at scale, if Southbound DB sends a port
related update before processing the transaction in which ovn-controller
claimed the logical port.

Reported-at: https://bugzilla.redhat.com/1946420
Fixes: 4d3cb42b076b ("binding: Set Logical_Switch_Port.up when all OVS flows are installed.")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>