Revert "lflow.c: Use warn level log for a lflow that has neither dp nor dpg."
This reverts commit 0ba2e1a2e2c413f057ac198e4c2bcf00ac40b19f.

System tests with "dp-groups=yes" are all failing due to the warning
log. The reason of the warning log (a lflow has neither dp nor dpg) is
because the dpg reference by the lflow is not monitored by the
ovn-controller yet when the lflow is firsly observed by the
ovn-controller. So dbg level log is valid with the current
implementation. There may be some inefficiency in such scenarios
because ovn-controller needs to be wakeup again when the dpg is
monitored later and the processing of the lflow would be delayed a
little. This may be addressed with future patches, if necessary.

Cc: Ilya Maximets <i.maximets@ovn.org>
Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>