ovn-controller: Fix incremental processing for multicast group dependency.
Multicast group changes are handled for physical flows in I-P, but not
handled for logical flows. Although logical flow doesn't care about the
content of a multicast group, the existance of it matters for logical
flow processing. If the multicast group is not found when the logical
flow is processed at first but laster it is created/monitored by
ovn-controller, the logical flow that has been processed before should
be reprocessed. This would avoid the workaound of ovn_lflow_add_unique()
used in northd when adding multicast group related logical flows.

Acked-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>