configure: Improve how to find ddlog binaries and libraries.
There was no easy way to use a ddlog installation from one of the
tarballs provided by the ddlog developers, because these put the
binaries in a subdirectory of an installation directory instead of in
a system- or user-level bin directory.  This commit makes that easier:
just do "--with-ddlog=$INSTALLDIR".

Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Acked-by: Mark Michelson <mmichels@redhat.com>