northd: Consolidate load balancer healthcheck/svc code.
Part of the load balancer healthcheck/svc configuration parsing was
done in ovn-northd.c and part in lib/lb.c.  Consolidate this in
ovn-northd.c.  As a side effect this will remove dependency on the
'ports' hmap in lib/lb.c.  This enables future commits to optimize
load balancer configuration parsing.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>