ovn-northd.c: Add proxy ARP support to OVN
This patch provides the ability to configure proxy ARP IPs on a Logical
Switch Router port. The IPs are added as Options for router ports. This
provides a useful feature where traffic for a service must be sent to an
address in a logical network address space, but the service is provided
in a different network. For example an NFS service is provide to Logical
networks at an address in their Logical network space, but the NFS
server resides in a physical network. A Logical switch Router port can
be configured to respond to ARP requests sent to the service "Logical
address", the Logical Router/Gateway can then be configured to forward
the traffic to the underlay/physical network.

(Note: northd-ddlog part is missing which needs to be added).

Signed-off-by: Brendan Doyle <brendan.doyle@oracle.com>
Signed-off-by: Numan Siddique <numans@ovn.org>